requirements.txt
numpy>=2.0.1
onnxruntime==1.19.0
opencv-python
pillow==10.2.0
torch>=2.4.0
torchvision>=0.19.0
tqdm==4.66.5
uniface==1.1.0
fastmcp>=2.0.0
